requirements.txt
streamlit==1.22.0
pillow==9.5.0
requests==2.28.2
replicate==0.8.3
python-dotenv==1.0.0
openai==0.27.8
matplotlib
opencv-python==4.5.5.64
reportlab==4.1.0 
